requirements.txt
Django>=4.2,<5
djangorestframework
xmltodict
psycopg2-binary
python-dotenv
celery[redis]
